requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
cloudinary==1.41.0
cloudinary-upload==0.0.4
distro==1.9.0
exceptiongroup==1.2.2
frozendict==2.3.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
moralis==0.1.49
numpy==2.2.0
openai==1.57.4
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pyTelegramBotAPI==4.25.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
